+++ src/Ingredients.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+function Ingredients({ingredients}) {
+  return (
+    <ul className="ingredients">
+      {ingredients.map((ingredient, i) => (
+        <li key={i}>{i}}: {ingredient.name}</li>
+      ))}
+    </ul>
+  );
+}
+
+export {Ingredients};
