+++ src/Menu.js
@@ -1,0 +1,20 @@
+import React from "react";
+import {Recipe} from './Recipe';
+
+
+function Menu({title, recipes}) {
+  return (
+    <article>
+      <header>
+        <h1>{title}</h1>
+      </header>
+      <div className="recipes">
+        {recipes.map((recipe, i) => (
+          <Recipe key={i} {...recipe} />
+        ))}
+      </div>
+    </article>
+  );
+}
+
+export {Menu};
