+++ src/Recipe.js
@@ -1,20 +1,13 @@
 import React from "react";
+import {Instructions} from './Instructions';
+import {Ingredients} from './Ingredients';
 
 function Recipe({ name, ingredients, steps }) {
   return (
     <section id={name.toLowerCase().replace(/ /g, "-")}>
       <h1>{name}</h1>
-      <ul className="ingredients">
-        {ingredients.map((ingredient, i) => (
-          <li key={i}>{ingredient.name}</li>
-        ))}
-      </ul>
-      <section className="instructions">
-        <h2>Cooking XXInstructions</h2>
-        {steps.map((step, i) => (
-          <p key={i}>{i}: {step}</p>
-        ))}
-      </section>
+      <Ingredients ingredients={ingredients} />
+      <Instructions title="Goodie foodie!" steps={steps} />
     </section>
   );
 }
